% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_powergrid_network_statistics.R
\name{create_powergrid_network_statistics}
\alias{create_powergrid_network_statistics}
\title{generate data for powergrid stats}
\usage{
create_powergrid_network_statistics(file_paths)
}
\arguments{
\item{file_paths}{string vector, the full file paths of the networks to analyse}
}
\description{
Produces a data frame that can be converted to a table for the paper
}
