% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_union_of_power_grids.R
\name{create_union_of_power_grids}
\alias{create_union_of_power_grids}
\title{put all the power grids into a single network}
\usage{
create_union_of_power_grids()
}
\description{
This is sugar to clean up the code and make a single powergrid with shared attributes from all the power grids
}
