% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validated_setse_embeddings.R
\name{cross_validated_setse_embeddings}
\alias{cross_validated_setse_embeddings}
\title{generate the dataframe showing the cross validated error of each metric}
\usage{
cross_validated_setse_embeddings(all_SETSe_emebeddings, PL_SETSe_emebeddings)
}
\arguments{
\item{all_SETSe_emebeddings}{A dataframe this is loaded earlier in the project}

\item{PL_SETSe_emebeddings}{A dataframe this is loaded earlier in the project}
}
\description{
This is used so that the error can be plotted showing the performance differences between the methods
elevation is excluded for now but can be re-included. This will require a change to the processing functions as well
}
