% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validated_stellar_embeddings.R
\name{cross_validated_stellar_embeddings}
\alias{cross_validated_stellar_embeddings}
\title{Produces performance metrics for the stellargraph embeddings}
\usage{
cross_validated_stellar_embeddings(stellar_all)
}
\arguments{
\item{stellar_all}{a dataframe.}
}
\description{
Takes a dataframe of stellargraph embeddings and returns a cross validated dataframe of
performance metrics from GAM across all dimensions with y being the point of loss of the giant component
}
