% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_robustness_plots.R
\name{generate_robustness_plots}
\alias{generate_robustness_plots}
\title{generate robustness plots}
\usage{
generate_robustness_plots(
  PL_SETSe_emebeddings,
  stellar_embeds,
  all_SETSe_emebeddings
)
}
\arguments{
\item{PL_SETSe_emebeddings}{dataframe}

\item{stellar_embeds}{dataframe}

\item{all_SETSe_emebeddings}{dataframe}
}
\description{
Creates the main plot and plots for the appendix
}
