% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_height_data.R
\name{generate_toy_height_data}
\alias{generate_toy_height_data}
\title{Function to clean up the toy heigh data chunk}
\usage{
generate_toy_height_data()
}
\description{
This function is just sugar to make the toy height data section easier to deal with
}
\details{
This function has no parameters and generates all results from simulated data.
The function outputs the toy_height_data dataframe that is used to plot the psychadelic beard which demonstrates
that the metic mean edge capacity is not great.
}
