% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{coerce values between 0 and 1}
\usage{
kappa(value)
}
\arguments{
\item{value}{a numeric vector. The vector of values to transform}
}
\description{
This function simply coerces the values of a vector to be a fraction of the minimum and maximum values
It is useful for putting everything on the same relative scale
}
