% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_holdout_results.R
\name{loess_holdout_results}
\alias{loess_holdout_results}
\title{cross validate performance using loes}
\usage{
loess_holdout_results(splits, model_dat, metrics_out = TRUE, ...)
}
\arguments{
\item{splits}{The random splits to be used when creating the model these are created by
vfold_cv from the rsample package}

\item{model_dat}{formula the formula that will be used to create the model}

\item{metrics_out}{logical output either the model performance metrics or the actual predictions}
}
\description{
cross validate performance using loes
}
