% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all_pl_setse_emebeddings.R
\name{process_all_pl_setse_emebeddings}
\alias{process_all_pl_setse_emebeddings}
\title{Process PL data}
\usage{
process_all_pl_setse_emebeddings(strain_folders, attack_folders)
}
\arguments{
\item{strain_folders}{A character string. The path to the strain folders}

\item{attack_folders}{A character string. The path to the attack folders}
}
\description{
This function takes the path of the PL attacks and the path of the PL embeddings and constructs the PL data frame
}
