% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all_setse_emebeddings.R
\name{process_all_setse_emebeddings}
\alias{process_all_setse_emebeddings}
\title{All concentrator Setse embeddings}
\usage{
process_all_setse_emebeddings(
  processed_path,
  folder_paths,
  graph_agg,
  PL_SETSe_emebeddings
)
}
\description{
Loads the setse data and saves it is not already saved

The function loads and processes the file output from the HPC script. It then saves the processed data
If the processed data is already saved this is loaded to save time.

@param processed_path A character string. The location of the processed files
@param folder_paths A character vector. The paths of each folder to load and process, each folder represents a graph
@param graph_agg A data frame of the summarised attack data that matches the strain data
@param PL_SETSe_emebeddings A dataframe the output of the pl equivalent of this function

@export
}
