% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all_setse_emebeddings2.R
\name{process_all_setse_emebeddings2}
\alias{process_all_setse_emebeddings2}
\title{All concentrator Setse embeddings}
\usage{
process_all_setse_emebeddings2(
  processed_path,
  folder_paths,
  graph_agg,
  PL_SETSe_emebeddings
)
}
\arguments{
\item{processed_path}{A character string. The location of the processed files}

\item{folder_paths}{A character vector. The paths of each folder to load and process, each folder represents a graph}

\item{graph_agg}{A data frame of the summarised attack data that matches the strain data}

\item{PL_SETSe_emebeddings}{A dataframe the output of the pl equivalent of this function}
}
\description{
Loads the setse data and saves it is not already saved

The function loads and processes the file output from the HPC script. It then saves the processed data
If the processed data is already saved this is loaded to save time.
}
