% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stellar_embeddings_loader_pl.R
\name{stellar_embeddings_loader_pl}
\alias{stellar_embeddings_loader_pl}
\title{load stellar embedding csv's}
\usage{
stellar_embeddings_loader_pl(folder_path)
}
\arguments{
\item{folder_path}{}
}
\description{
This function loads the csvs that are created when the embeddings are made by the stellar python library.
The function assumes that the graphs were created from graphml files. The scripts that created the embeddings
This function is just for the proportionally loaded files
}
